moved piston theory test into framework
