+++ tests/framework/test_piston_framework.py
@@ -1,4 +1,4 @@
-import numpy as np
+import numpy as np, unittest
 from mpi4py import MPI
 from funtofem import TransferScheme
 
@@ -10,10 +10,7 @@ from pyfuntofem.interface import (
     CommManager,
 )
 from pyfuntofem.driver import FUNtoFEMnlbgs, TransferSettings
-
-from structural_model import OneraPlate
-import unittest
-
+from tacs import TACS, elements, functions, constitutive
 
 class CoupledFrameworkTest(unittest.TestCase):
     def _setup_model_and_driver(self):
@@ -212,6 +209,87 @@ class CoupledFrameworkTest(unittest.TestCase):
             # assert pass_
 
         return
+
+def OneraPlate(tacs_comm):
+    # Set the creator object
+    ndof = 6
+    creator = TACS.Creator(tacs_comm, ndof)
+
+    if tacs_comm.rank == 0:
+        # Create the elements
+        nx = 10
+        ny = 10
+
+        # Set the nodes
+        nnodes = (nx + 1) * (ny + 1)
+        nelems = nx * ny
+        nodes = np.arange(nnodes).reshape((nx + 1, ny + 1))
+
+        conn = []
+        for j in range(ny):
+            for i in range(nx):
+                # Append the node locations
+                conn.append(
+                    [nodes[i, j], nodes[i + 1, j], nodes[i, j + 1], nodes[i + 1, j + 1]]
+                )
+
+        # Set the node pointers
+        conn = np.array(conn, dtype=np.intc).flatten()
+        ptr = np.arange(0, 4 * nelems + 1, 4, dtype=np.intc)
+        elem_ids = np.zeros(nelems, dtype=np.intc)
+        creator.setGlobalConnectivity(nnodes, ptr, conn, elem_ids)
+
+        # Set the boundary conditions - fixed on the root
+        bcnodes = np.array(nodes[:, 0], dtype=np.intc)
+        creator.setBoundaryConditions(bcnodes)
+
+        root_chord = 0.8
+        semi_span = 1.2
+        taper_ratio = 0.56
+        sweep = 26.7  # degrees
+
+        # Set the node locations
+        Xpts = np.zeros(3 * nnodes, TACS.dtype)
+        x = np.linspace(0, 1, nx + 1)
+        y = np.linspace(0, 1, ny + 1)
+        for j in range(ny + 1):
+            for i in range(nx + 1):
+                c = root_chord * (1.0 - y[j]) + root_chord * taper_ratio * y[j]
+                xoff = 0.25 * root_chord + semi_span * y[j] * np.tan(
+                    sweep * np.pi / 180.0
+                )
+
+                Xpts[3 * nodes[i, j]] = xoff + c * (x[i] - 0.25)
+                Xpts[3 * nodes[i, j] + 1] = semi_span * y[j]
+
+        # Set the node locations
+        creator.setNodes(Xpts)
+
+    # Set the material properties
+    props = constitutive.MaterialProperties(rho=2570.0, E=70e9, nu=0.3, ys=350e6)
+
+    # Set constitutive properties
+    t = 0.025
+    tnum = 0
+    maxt = 0.015
+    mint = 0.015
+    con = constitutive.IsoShellConstitutive(props, t=t, tNum=tnum)
+
+    # Create a transformation object
+    transform = elements.ShellNaturalTransform()
+
+    # Create the element object
+    element = elements.Quad4Shell(transform, con)
+
+    # Set the elements
+    elems = [element]
+    creator.setElements(elems)
+
+    # Create TACS Assembler object from the mesh loader
+    assembler = creator.createTACS()
+
+    return assembler
+
 
 
 if __name__ == "__main__":
